refactoring handlers

+++ src/Application/TrackerUsersComparison/Queries/GetUserComparisonDto.cs
@@ -6,7 +6,7 @@ public class GetUserComparisonDto
 {
     public int Id { get; set; }
     public string? Name { get; set; }
-    public decimal TotalHours { get; set; }
+    public double TotalHours { get; set; }
 
     private class Mapping : Profile
     {
